move PartialUser to events module
