+++ src/models/commands.rs
@@ -1,4 +1,4 @@
-use super::shared::PartialUser;
+use super::events::PartialUser;
 
 builder! {SubscriptionArgs
     secret: String,     // Activity{Join,Spectate}
