--- src/models/shared.rs
@@ -1,6 +1,0 @@
-builder! {PartialUser
-    id:            String,
-    username:      String,
-    discriminator: String,
-    avatar:        String,
-}
